Remove component composition (#51)
* remove component composition

* remove styling

* rename directory to be consistent with other names
